tfw xorg failed on me

+++ .bashrc
@@ -1,18 +1,28 @@
 # Mirko van der Waal
-# Set default shell values.
-export LANG="en_GB.UTF-8"
+# <Mvdw at airmail dot cc>
 
-# export PS1="\[\e[00;31m\]\u\[\e[0m\]\[\e[00;37m\]@\[\e[0m\]\[\e[00;32m\]\h\[\e[0m\]\[\e[00;37m\] [\[\e[0m\]\[\e[00;33m\]\w\[\e[0m\]\[\e[00;37m\]] \[\e[0m\]"
-export PS1="\[\e[00;34m\]\w\[\e[0m\]\[\e[00;37m\] \[\e[0m\]\[\e[00;34m\]>>>\[\e[0m\]\[\e[00;37m\] \[\e[0m\]"
+# If not running interactively, don't do anything
+[[ $- != *i* ]] && return
 
-# Connect to the VPS, currently only have one.
+# This is already set when `locale-gen` was ran.
+# Exporting it again doesn't hurt.
+export LANG="en_US.UTF-8"
+# Other "very useful" exports.
+export PS1="\[\e[00;34m\]\w\[\e[0m\]\[\e[00;37m\] \[\e[0m\]\[\e[00;34m\]%\[\e[0m\]\[\e[00;37m\] \[\e[0m\]"
+
+export EDITOR="vim"
+export BROWSER="firefox"
+
+# Quickly connect to my VPS server, 
+# NOTE: Learn how to use a VPN.
 alias vps="ssh mvdw@128.199.46.104"
 
 # Merge the .Xresources file after we worked with it.
 alias xmerge="xrdb -merge ~/.Xresources"
 
-# Rebind ls to show colors and premissions.
 alias ls="ls -aG --color=auto"
+
+alias ix="curl -F 'f:1=<-' ix.io"
 
 # Kill current shell (Leaves no trace of commands.)
 alias kcs="kill -9 $$"		
@@ -21,6 +31,6 @@ alias kcs="kill -9 $$"
 alias sudo="sudo "
 
 # Get your outer IP address.
-alias ip="curl icanhazip.com"
+alias ipaddr="curl icanhazip.com"
+alias lpaddr="ifconfig"
 
-alias maze-smooth='Z=╱╲;for((;;)){ printf ${Z:RANDOM&1:1};sleep .01;}'
